examples/am65x: Cleanup Halt and MAC_Multiply_Accum Makefiles
Replace the usage of a Makefile variable SUBSYS with the more
appropriate SLICE in the PRU_Halt, RTU_Halt, PRU_MAC_Multiply_Accum
and RTU_MAC_Multiply_Accum examples. The SUBSYS was primarily
meant to mean the PRUSS instance, while SLICE represents the
correct partition within a PRUSS the PRU or RTU core belongs to.

There is no functional difference in the outputs.

Signed-off-by: Suman Anna <s-anna@ti.com>